moved manage profile from auth components to pages
